rename some classes/methods for clarity and move constants to util package
