+++ modules/rabbitmq/src/main/java/org/apache/axis2/transport/rabbitmq/utils/RabbitMQUtils.java
@@ -29,7 +29,6 @@ import org.apache.axis2.builder.BuilderUtil;
 import org.apache.axis2.builder.SOAPBuilder;
 import org.apache.axis2.context.MessageContext;
 import org.apache.axis2.transport.TransportUtils;
-import org.apache.axis2.transport.rabbitmq.RabbitMQConstants;
 import org.apache.axis2.transport.rabbitmq.RabbitMQMessage;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
